tests: Split out s3 and local tests.
This mirrors the split of the code in 7c0d414aff10.
